confirm
confirm unfollow

+++ reactproject/frontend/src/components/pages/PagesUser/HomeLayout/Center.jsx
@@ -167,6 +167,13 @@ const Center = () => {
 
     try {
       if (isFollowing) {
+        const confirmed = window.confirm(
+          `Are you sure you want to unfollow this ${followUserId}?`
+        );
+
+        if (!confirmed) {
+          return;
+        }
         await axios.delete(`http://localhost:8081/unfollow/${followUserId}`, {
           data: { followerId: user.userID },
         });
